Rename heyKeyId -> openPgpKeyId
